First Draft
This app is designed to integrate data from instruments T4S2480 and
T4S2481. As long as they are correctly named, then the empirical
calibration for each will be applied. A manual override is also
available.

Data is also converted to molecular form.

The empirical quantifications were altered to handle NA data - an
improvement I may propagate to other apps.